Update format of CDN log line
We've recently changed these log lines to include the name of the backend
where the response came from. Some examples of expected backends are 'origin',
'mirror0' or 'mirror1' though there may be others.

This commit also updates the test fixtures and fixes the code which
has broken as a result.