Add script to store stats on CDN backends
We want to be able to see which backends are being used to serve requests
for GOV.UK.

In order to reduce the amount of stuff stored in Graphite we've not
included the 'origin' backend as part of the `VALID_CDN_FAILOVERS`
list. Origin should serve almost all the traffic but really we're
interested in the bits which aren't being served by origin.