Documentation and design for new approach
This contains a detailed description of how the log processing and
monitoring works, in the redesigned system.

This is an attempt to address several problems with the existing mix of shell
and ruby scripts: firstly, the mix of shell and ruby is a bit hard to follow.

Secondly, the log files aren't actually for a single day, and sometimes we have
several days logs in a single file, so we need to parse the timestamps in the
file - which includes being able to parse timezone information, and then merge
the resulting counts from different files to get the correct counts for a
single day.

Thirdly, we want to be easily able to ensure that we have accurate counts for
the log files which exist, even if they get modified (or changed to a
compressed form) - so this design moves to having a single script for counting
all the log files which are present, which will skip log files which have
already been counted: it ensures that a log file which is later modified is
re-counted by tracking the file size when the log was counted.  This script
will be run nightly to update the derived counts.

Finally, and perhaps most importantly, we may not be able to keep the GOV.UK
logs indefinitely, due to data size and potential privacy implications, but a
digest of just the counts of accesses to URLs in a time period will take up
much less space, have fewer privacy risks, and be almost as useful for data
analysis. We expect in future to want to do more than just look for successful
requests in the logs, so this commit takes the approach of first running a pure
count of the combination of hour of day, HTTP method, path, and resulting
status code.

A later stages of processing are then done to using the counts as input,
producing derived information such as lists of successfully accessed pages per
day.  This processing can be adjusted as needed, without the need to refer back
to the original source log files (which can then be deleted if needed without
affecting the future options for analysis).