Separate out file streaming from log parsing
Ruby's internal GzipReader appears to be incapable of processing large
files without buffering the whole output in memory.  Therefore, use an
external call to `gunzip` instead.

In order to do this, we use the same LogStreamer code which we're
already using to stream output from an external process:

 - we rename the existing LogStreamer class to LogTailStreamer, which is
   a more descriptive name (it only tails the end of the log file,
   ignoring existing bits of the file).

 - we refactor the implementation of LogTailStreamer to separate out the
   definition of the process it calls, and abstract streaming the output
   of the process to a ProcessStreamer class.

 - we replace the LogParser::file_reader class method with a
   LogFileStreamer class

To ensure the `gunzip` process is cleaned up properly, LogFileStreamer
takes a block and calls it with the stream from `gunzip`.  This means we
can't continue to support the `LogParser.open` method with the same
semantics (returning a parser with an open stream), and this is a bit
non standard anyway, so we just remove it and call it explicitly.