Use external sort and uniq
Counting a day's accesses in memory was using well over 1Gb of memory,
and getting very slow as a result.  Instead, use the UNIX sort command
to sort the accesses into order, then use `uniq -c` to count the
accesses.  These tools are already optimised to work with large files.
I've tested this on some large datasets, processing them in a few
minutes.