Don't consider 3xx status codes to be "good"
When calculating lists of urls which return successful status codes,
we're getting a very large number of urls which are redirects.  Some
of these redirects are even to other urls which then return a 404 (eg,
stripping a trailing slash from the URL).

To reduce the volume of URLs returned, don't consider 3xx status codes
as "good".