Fix bug in sorting of first-last-success file
The sort command for creating the "first-last-success" file was
incorrectly using the system locale, rather than setting the C locale.
This meant that URLs such as `/a-url` and `//a-url` were being compared
as equal, and thus each URL wasn't being grouped together appropriately.

This means that the known-good URLs file has lots of missing entries.
It also makes the sort command significantly slower (because it has to
perform a much more complicated comparison than simple byte-wise
comparison).

After this is deployed, the generated "first-last-successes" file needs
to be deleted, so that the next run will generate it correctly from
scratch.