Only log known-good failures for GET requests
The list of "known good" URLs is generated from GET requests.
Therefore, it will often be an error if a request using another method
(such as POST or OPTIONS) is made to these URLs.

We want the "known good" failure list to be noise free, so that we can
trust that entries in it are things that we should be alerting on.
Therefore, we'll ignore log entries the aren't from GET requests when
checking for known good failures.