Log time of the last successful access in Kibana
We want to record the time of the last successful access in Kibana, so
that we can generate a report of URLs which have recently started to
fail (separating this of URL which have frequently failed), and also to
emit to statsd information on the number of URLs which have recently
started failing.

This commit adds the date of the last successful access to each line in
the list of known_good URLs.

It then reads that date in when parsing the file, and stores the list of
known_good URLs in a hash instead of an array, associating the date to
the URL.

It uses this data to add an extra field to the logstash output recording
the date when the URL last worked (this should help to identify changes
which caused URLs to stop working) and adds a tag of
"recent_known_good_fail" to such logstash messages.

It also emits separate statsd output for failures of URLs which have had
a successful request in the last 7 days.

This also stops sending a dummy "length" field with a value of "-" to
logstash; this data isn't available, but it's fine to omit it.