Upgrade to ruby 2.4.2
We've removed ruby 2.2.3 from our servers.  Version 2.2.8 is available
but we'd like to remove it eventually too so going up to 2.4.2 should
give us some headroom before the same problem occurs again.

The one change we need to make is to explicitly require tmpdir in order
to get the `mktmpdir` method available on `Dir` for our specs.